More IronPython bytes compatibility issues. Fixes Digital-Structures/ghpythonremote#6
IronPython 2.7.0 cannot do b"".join(["a"]), so brine needs to
absolutely always dump bytes into the stream (in the `brine.dump()`
method). The `_dump_int` method was returning strings on immutable
integers, so fix that.

The BYTES_LITTERAL method is not really working for IronPython, probably
should be changed but also needs to be split because some usages really
need a str output.